requirements.txt
fastapi
uvicorn
pydantic
boto3
psycopg2-binary
pymongo
pytesseract
Pillow
sentence-transformers
transformers
datasets
pyyaml
python-docx
